Organize more, add todo
